Add commands to sending

+++ app/src/main/java/ru/art241111/graphicaltoolforkawasaki/repository/enity/Commands.kt
@@ -4,6 +4,6 @@ import ru.art241111.graphicaltoolforkawasaki.repository.enity.enums.Coordinate
 import ru.art241111.graphicaltoolforkawasaki.repository.enity.enums.TypesOfMovementToThePoint
 
 data class Move(val coordinate: Coordinate, val sizeOfPlant: Int):RobotCommands
-data class MoveToPoint(val type: TypesOfMovementToThePoint, val coordinate: List<Int>):RobotCommands
+data class MoveToPoint(val type: TypesOfMovementToThePoint,val coordinate: List<Float>):RobotCommands
 class OpenGripper :RobotCommands
 class CloseGripper:RobotCommands
